creating and edititng articles

+++ blog/app/Http/Controllers/ArticlesController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Article;
-use \Illuminate\Support\Facades\Input;
+use Illuminate\Support\Facades\Input;
+use App\Http\Requests\ArticleRequest;
+use Illuminate\Support\Facades\Auth;
+
 class ArticlesController extends Controller
 {
+    function __construct() {
+        $this->middleware('auth',['except' => ['index']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,9 +20,9 @@ class ArticlesController extends Controller
      */
     public function index()
     {
-       $articles = Article::all();
-       
-       return view('articles.list',['articles' =>$articles]);
+        $articles = Article::all();
+//        compact('articles');
+        return view('articles.list',['articles' => $articles]);
     }
 
     /**
@@ -26,9 +32,8 @@ class ArticlesController extends Controller
      */
     public function create()
     {
-       $categories = \App\Category::pluck('title','id');
-       
-        return view('articles.create',['categories' => $categories]);
+        $categories = \App\Category::pluck('title', 'id');
+        return view('articles.create', ['categories' => $categories]);
     }
 
     /**
@@ -37,17 +42,32 @@ class ArticlesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ArticleRequest $request)
     {
         
+//        $data = $request->all();
+//        $data['user_id'] = Auth::user()->id;
+//        $image = Input::file('image');
+//        if(isset($image)) {
+//            $imageName = time().'.'.$image->getClientOriginalExtension();
+//            $data['image'] = $imageName;
+//            $image->move("uploads",$imageName);
+//        } else {
+//            $data['image'] = '';
+//        }
+//
+//        Article::create($data);
+        
         $data = $request->all();
-        $data['user_id'] = 1;
         $image = Input::file('image');
-        $imageName = time().'.'.$image->getClientOriginalExtension();
-        $data['image'] = $imageName;
-        $image->move("uploads",$imageName);
-        //var_dump($imageName);exit;
-        Article::create($data);
+        if(isset($image)) {
+            $imageName = time().'.'.$image->getClientOriginalExtension();
+            $data['image'] = $imageName;
+            $image->move("uploads",$imageName);
+        } else {
+            $data['image'] = '';
+        }
+        Auth::user()->articles()->save(new Article($data));
         return redirect('article');
     }
 
@@ -59,11 +79,11 @@ class ArticlesController extends Controller
      */
     public function show($slug)
     {
-       $article = Article::where(['slug' => $slug])->first();
-    if(!isset($article)){
-        return abort("404");
-    }
-       return view('articles.show', ['article' => $article]);
+        $article = Article::where(['slug' => $slug])->first();
+        if(!isset($article)) {
+            return abort("404");
+        }
+        return view('articles.show', ['article' => $article]);
     }
 
     /**
@@ -75,10 +95,8 @@ class ArticlesController extends Controller
     public function edit($id)
     {
         $article = Article::findOrFail($id);
-        
-        $categories = \App\Category::pluck('title','id');
-       
-        return view('articles.edit',['categories' => $categories, 'article' => $article]);
+        $categories = \App\Category::pluck('title', 'id');
+        return view('articles.edit', ['categories' => $categories, 'article' => $article]);
     }
 
     /**
@@ -88,22 +106,23 @@ class ArticlesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(ArticleRequest $request, $id)
     {
         $article = Article::findOrFail($id);
         $data = $request->all();
-        $data['user_id'] = 1;
+        $data['user_id'] = Auth::user()->id;
         $image = Input::file('image');
-        $imageName = time().'.'.$image->getClientOriginalExtension();
-        $data['image'] = $imageName;
-        $image->move("uploads",$imageName);
-        
-        //delete old img
-        if(file_exists('uploads/'.$article->image)){
-            unlink('uploads/'.$article->image);
+        if(isset($image)) {
+            $imageName = time().'.'.$image->getClientOriginalExtension();
+            $data['image'] = $imageName;
+            $image->move("uploads",$imageName);
+            
+             //delete old img
+            if($article->image != '' && file_exists("uploads/".$article->image)) {
+                unlink("uploads/".$article->image);
+            }
         }
         $article->update($data);
-//        var_dump($data);exit;
         return redirect('article');
     }
